refactor user commands to own module
