+++ base/src/lib.rs
@@ -23,6 +23,7 @@ pub mod protocol_state_proof;
 pub mod protocol_version;
 mod serialization_type_conversions;
 pub mod staged_ledger_diff;
+pub mod user_commands;
 pub mod verification_key;
 
 /// Re-export all the public types under this module for convenience
@@ -38,5 +39,6 @@ pub mod types {
     pub use super::protocol_state_proof::*;
     pub use super::protocol_version::*;
     pub use super::staged_ledger_diff::*;
+    pub use super::user_commands::*;
     pub use super::verification_key::*;
 }
