+++ base/src/staged_ledger_diff.rs
@@ -8,12 +8,11 @@
 #![allow(missing_docs)]
 
 use crate::types::TokenId;
-use derive_more::From;
-use mina_crypto::signature::{PublicKey, Signature};
+use mina_crypto::signature::PublicKey;
 use serde::{Deserialize, Serialize};
-use thiserror::Error;
 
-use crate::numbers::{Amount, ExtendedU32};
+use crate::numbers::Amount;
+use crate::user_commands::UserCommand;
 
 #[derive(Clone, Serialize, Deserialize, Default, PartialEq, Debug)]
 /// Top level wrapper type for a StagedLedgerDiff
@@ -54,108 +53,6 @@ pub struct UserCommandWithStatus {
     pub data: UserCommand,
     pub status: TransactionStatus,
 }
-
-#[derive(Clone, Serialize, Deserialize, PartialEq, Debug)]
-#[non_exhaustive]
-pub enum UserCommand {
-    SignedCommand(SignedCommand),
-    // FIXME: other variants are not covered by current test block
-}
-
-impl Default for UserCommand {
-    fn default() -> Self {
-        Self::SignedCommand(SignedCommand::default())
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize, Default, PartialEq, Debug)]
-pub struct SignedCommand {
-    pub payload: SignedCommandPayload,
-    pub signer: PublicKey,
-    pub signature: Signature,
-}
-
-#[derive(Clone, Serialize, Deserialize, Default, PartialEq, Debug)]
-pub struct SignedCommandPayload {
-    pub common: SignedCommandPayloadCommon,
-    pub body: SignedCommandPayloadBody,
-}
-
-#[derive(Clone, Serialize, Deserialize, Default, PartialEq, Debug)]
-pub struct SignedCommandPayloadCommon {
-    pub fee: Amount,
-    pub fee_token: TokenId,
-    pub fee_payer_pk: PublicKey,
-    pub nonce: ExtendedU32,
-    pub valid_until: ExtendedU32,
-    pub memo: SignedCommandMemo,
-}
-
-#[derive(Clone, Serialize, Deserialize, PartialEq, Debug)]
-#[non_exhaustive]
-pub enum SignedCommandPayloadBody {
-    PaymentPayload(PaymentPayload),
-    // FIXME: other variants are not covered by current test block
-}
-
-impl Default for SignedCommandPayloadBody {
-    fn default() -> Self {
-        Self::PaymentPayload(PaymentPayload::default())
-    }
-}
-
-#[derive(Clone, Serialize, Deserialize, Default, PartialEq, Debug)]
-pub struct PaymentPayload {
-    pub source_pk: PublicKey,
-    pub receiver_pk: PublicKey,
-    pub token_id: TokenId,
-    pub amount: Amount,
-}
-
-#[derive(Clone, Serialize, Deserialize, Default, PartialEq, Debug, From)]
-pub struct SignedCommandMemo(pub Vec<u8>);
-
-impl TryFrom<&str> for SignedCommandMemo {
-    type Error = SignedCommandMemoError;
-    fn try_from(s: &str) -> Result<Self, Self::Error> {
-        const DIGEST_LEN: usize = 32;
-        const MAX_INPUT_STRING_LENGTH: usize = DIGEST_LEN;
-        const MEMO_LEN: usize = DIGEST_LEN + 2;
-        const TAG_INDEX: usize = 0;
-        // const DIGEST_TAG: u8 = 0;
-        const LEN_INDEX: usize = 1;
-        const BYTES_TAG: u8 = 1;
-        if s.len() > MAX_INPUT_STRING_LENGTH {
-            return Err(SignedCommandMemoError::StringTooLong);
-        }
-        let mut v = vec![0; MEMO_LEN];
-        v[TAG_INDEX] = BYTES_TAG;
-        v[LEN_INDEX] = s.len() as u8;
-        for (i, b) in s.as_bytes().iter().enumerate() {
-            v[i + 2] = *b;
-        }
-        Ok(Self(v))
-    }
-}
-
-impl TryFrom<String> for SignedCommandMemo {
-    type Error = SignedCommandMemoError;
-    fn try_from(s: String) -> Result<Self, Self::Error> {
-        Self::try_from(s.as_str())
-    }
-}
-
-// TODO impl Into<String> for SignedCommandMemo
-
-#[derive(Debug, Error)]
-pub enum SignedCommandMemoError {
-    #[error("Input string is too long")]
-    StringTooLong,
-}
-
-#[derive(Clone, Serialize, Deserialize, Default, PartialEq, Debug)]
-// FIXME: No test coverage yet
-pub struct SnappCommand;
 
 #[derive(Clone, Serialize, Deserialize, PartialEq, Debug)]
 #[non_exhaustive]
